+++ src/Presentation/SmartStore.Web/Controllers/CatalogHelper.cs
@@ -170,7 +170,7 @@ namespace SmartStore.Web.Controllers
 				ProductType = product.ProductType,
 				VisibleIndividually = product.VisibleIndividually,
 				//Manufacturers = _manufacturerService.GetProductManufacturersByProductId(product.Id),
-				Manufacturers = PrepareManufacturersOverviewModel(_manufacturerService.GetProductManufacturersByProductId(product.Id)),
+				Manufacturers = PrepareManufacturersOverviewModel(_manufacturerService.GetProductManufacturersByProductId(product.Id), null, true),
 				ReviewCount = product.ApprovedTotalReviews,
 				DisplayAdminLink = _services.Permissions.Authorize(StandardPermissionProvider.AccessAdminPanel),
 				//EnableHtmlTextCollapser = Convert.ToBoolean(_settingService.GetSettingByKey<string>("CatalogSettings.EnableHtmlTextCollapser")),
@@ -1392,7 +1392,7 @@ namespace SmartStore.Web.Controllers
 
 				if (prepareManufacturers)
 				{
-					model.Manufacturers = PrepareManufacturersOverviewModel(_manufacturerService.GetProductManufacturersByProductId(product.Id), cachedManufacturerModels);
+					model.Manufacturers = PrepareManufacturersOverviewModel(_manufacturerService.GetProductManufacturersByProductId(product.Id), cachedManufacturerModels, false);
 				}
 
 				if (_catalogSettings.ShowBasePriceInProductLists)
@@ -1723,7 +1723,8 @@ namespace SmartStore.Web.Controllers
 
 		public List<ManufacturerOverviewModel> PrepareManufacturersOverviewModel(
 			ICollection<ProductManufacturer> manufacturers, 
-			IDictionary<int, ManufacturerOverviewModel> cachedModels = null)
+			IDictionary<int, ManufacturerOverviewModel> cachedModels = null,
+			bool forProductDetailPage = false)
 		{
 			var model = new List<ManufacturerOverviewModel>();
 
@@ -1748,7 +1749,10 @@ namespace SmartStore.Web.Controllers
 
 					};
 
-                    item.PictureModel = PrepareManufacturerPictureModel(manufacturer, manufacturer.GetLocalized(x => x.Name));
+					if (_catalogSettings.ShowManufacturerPicturesInProductDetail)
+					{
+						item.PictureModel = PrepareManufacturerPictureModel(manufacturer, manufacturer.GetLocalized(x => x.Name));
+					}
 
 					cachedModels.Add(item.Id, item);
 				}
@@ -1763,11 +1767,11 @@ namespace SmartStore.Web.Controllers
         {
             var model = new PictureModel();
 
-            int pictureSize = _mediaSettings.ManufacturerThumbPictureSize;
+            var pictureSize = _mediaSettings.ManufacturerThumbPictureSize;
             var manufacturerPictureCacheKey = string.Format(ModelCacheEventConsumer.MANUFACTURER_PICTURE_MODEL_KEY,
                 manufacturer.Id,
                 pictureSize,
-                true,
+				!_catalogSettings.HideManufacturerDefaultPictures,
                 _services.WorkContext.WorkingLanguage.Id,
                 _services.WebHelper.IsCurrentConnectionSecured(),
                 _services.StoreContext.CurrentStore.Id);
@@ -1778,7 +1782,7 @@ namespace SmartStore.Web.Controllers
                 {
                     PictureId = manufacturer.PictureId.GetValueOrDefault(),
                     //FullSizeImageUrl = _pictureService.GetPictureUrl(manufacturer.PictureId.GetValueOrDefault()),
-                    ImageUrl = _pictureService.GetPictureUrl(manufacturer.PictureId.GetValueOrDefault(), pictureSize),
+                    ImageUrl = _pictureService.GetPictureUrl(manufacturer.PictureId.GetValueOrDefault(), pictureSize, !_catalogSettings.HideManufacturerDefaultPictures),
                     Title = string.Format(T("Media.Manufacturer.ImageLinkTitleFormat"), localizedName),
                     AlternateText = string.Format(T("Media.Manufacturer.ImageAlternateTextFormat"), localizedName)
                 };